+++ src/routes/RequestsRoute.js
@@ -65,6 +65,7 @@ import {
   DEFAULT_REQUEST_TYPE_VALUE,
   INPUT_REQUEST_SEARCH_SELECTOR,
   PRINT_DETAILS_COLUMNS,
+  requestFilterTypes,
 } from '../constants';
 import {
   buildUrl,
@@ -677,7 +678,7 @@ class RequestsRoute extends React.Component {
     this.setCurrentServicePointId();
   }
 
-  componentDidUpdate(prevProps, prevState) {
+  componentDidUpdate(prevProps) {
     const patronBlocks = get(this.props.resources, ['patronBlocks', 'records'], []);
     const prevBlocks = get(prevProps.resources, ['patronBlocks', 'records'], []);
     const { submitting, isViewPrintDetailsEnabled } = this.state;
@@ -729,13 +730,30 @@ class RequestsRoute extends React.Component {
       this.onSearchComplete(resources.records);
     }
 
-    if (isViewPrintDetailsEnabled !== prevState.isViewPrintDetailsEnabled && !isViewPrintDetailsEnabled) {
-      this.columnHeadersMap = getFilteredColumnHeadersMap(this.columnHeadersMap);
+    if (!isViewPrintDetailsEnabled) {
+      this.handlePrintDetailsDisabled();
+    }
+  }
+
+  handlePrintDetailsDisabled() {
+    /**
+     * The function handles the following actions when `isViewPrintDetailsEnabled` is false:
+     *
+     * 1. If `filters` in query includes 'PRINT STATUS' filter:
+     *    - it clears the 'PRINT STATUS' filter from query by invoking `handleFilterChange`.
+     *
+     * 2. If `sort` in query includes sort by 'printed' or 'copies' column:
+     *    - it removes sorting by 'printed' or 'copies' from the sort query. If both are being sorted,
+     *    the sorting is updated to 'requestDate' instead.
+    */
+    const { resources: { query }, mutator } = this.props;
+    const printStatusFilterInQuery = this.getActiveFilters()[requestFilterTypes.PRINT_STATUS];
+
+    if (printStatusFilterInQuery?.length) {
+      this.handleFilterChange({ name: requestFilterTypes.PRINT_STATUS, values: [] });
     }
 
-    if (!isViewPrintDetailsEnabled && (query.sort?.includes('printed') || query.sort?.includes('copies'))) {
-      // Remove 'copies' and 'printed' from query sorting when the user disables
-      // 'Enable view print details (Pick slips)' in settings and returns to the Requests App.
+    if (query.sort?.includes('printed') || query.sort?.includes('copies')) {
       const sort = updateQuerySortString(query.sort);
       mutator.query.update({ sort });
     }
@@ -833,6 +851,9 @@ class RequestsRoute extends React.Component {
     queryString = queryClauses.join(' and ');
     const records = await this.fetchReportData(this.props.mutator.reportRecords, queryString);
     const recordsToCSV = this.buildRecords(records);
+
+    this.columnHeadersMap = this.state.isViewPrintDetailsEnabled ? this.columnHeadersMap :
+      getFilteredColumnHeadersMap(this.columnHeadersMap);
 
     exportCsv(recordsToCSV, {
       onlyFields: this.columnHeadersMap,