+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Optimize performance of the "print-events-entry" API to reduce slowness during the initial call. Refs UIREQ-1130.
 * Revert Custom "Print Status" filter code implementation. Refs UIREQ-1146.
 * Add sorting to 'Printed' and '# Copies' columns in the Request App. Refs UIREQ-1140.
+* Implement "Print Status" filters as server-side filters. Refs UIREQ-1147.
 
 ## [9.1.2] (https://github.com/folio-org/ui-requests/tree/v9.1.2) (2024-09-13)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v9.1.1...v9.1.2)